add the initial implementation and tests
This is the first-pass implementation and handling. This adds a
`railtie.rb` for inserting into Rails's boot/initializer flow.
This also adds a TemplateHandler for Rails. Finally, this adds
some default configuration to setup using Nm in Rails. Otherwise,
things should just work as expected.

This brings in the latest version of Nm that has been rewritten
specifically for this gem. It optionally runs templates against
an injeted context. This takes the ActionView::Base instance from
rails and runs the Nm templates in that. This means you get all of
Rails's template helpers and instance variables.

I chose to implement a system test that rendered a template that
used all of Nm's API: `node`, `map`, and `partial` for coverage.

Note: I got a lot of inspiration from looking at e.g. jbuilder and
slim's rails template handlers. I also got inspiration on setting
up the test helpers from jbuilder.